move `drain_fulfillment_cx_or_panic` to be private to traits::trans
